Key/requirements.txt
qrcode
pillow
pyzbar
opencv-python
numpy
django
django-cors-header
djongo
djangorestframework
typing_extensions
tzdata
Django
